requirements.txt
Authlib==0.14.3
Flask==2.0.2
Flask-CORS
Jinja2==3.0.3
Werkzeug==2.0.2
requests
python-dotenv
python-dateutil
